requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
dnspython==2.2.1
eventlet==0.33.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
